AdminDAO and UtenteRegistratoDAO created. Javadoc added for all the classes who were missing it and codestyle details updated.
