--- src/main/java/C15/dev/HeartCareApplication.java
@@ -1,13 +1,0 @@
-package C15.dev;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class HeartCareApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(HeartCareApplication.class, args);
-	}
-
-}
