--- src/main/java/C15/dev/model/entity/Admin.java
@@ -1,15 +1,0 @@
-package C15.dev.model.entity;
-
-import jakarta.persistence.*;
-
-
-/**
- * @author Leopoldo Todisco
- * Creato il: 29/12/2022
- * Questa è la classe relativa a un Admin.
- */
-@Entity
-public class Admin extends UtenteRegistrato{
-
-
-}
