+++ src/main/java/c15/dev/HeartCareApplication.java
@@ -1,0 +1,13 @@
+package c15.dev;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class HeartCareApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(HeartCareApplication.class, args);
+	}
+
+}
