+++ src/main/java/c15/dev/model/dao/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * package che contiene tutti i DAO del sistema HeartCare.
+ */
+package c15.dev.model.dao;
