+++ src/main/java/c15/dev/model/entity/enumeration/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * in questo package si trovano le enumerazioni utilizzate
+ */
+package c15.dev.model.entity.enumeration;
