+++ src/main/java/c15/dev/model/entity/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * package che contiene tutte le entità del sistema HeartCare.
+ */
+package c15.dev.model.entity;
