+++ src/main/resources/application.properties
@@ -1,7 +1,7 @@
 spring.jpa.hibernate.ddl-auto=update
 spring.datasource.url=jdbc:mysql://${MYSQL_HOST:localhost}:3306/heartcare
 spring.datasource.username=root
-spring.datasource.password=Ciao123!
+spring.datasource.password=123leo456
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.jpa.show-sql=true
