--- src/test/java/C15/dev/HeartCareApplicationTests.java
@@ -1,13 +1,0 @@
-package C15.dev;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class HeartCareApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
