+++ src/test/java/c15/dev/HeartCareApplicationTests.java
@@ -1,0 +1,13 @@
+package c15.dev;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class HeartCareApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
